FEATRUE: pin windows
Pressing F2 on any entry of the list will pin the entry.
This means the entry will show up in the list even if its not running.
It will show at the bottom of the list in grey color.
Activating on of those grey entries while the task is not running will launch the task.
If the task is already running it will show in the regular list and will not be shown in grey at the bottom.
All pinned tasks have a "P" in the column on the left end of the list.